catch empty blob

+++ parsons/google/google_cloud_storage.py
@@ -318,8 +318,10 @@ class GoogleCloudStorage:
         """
 
         blob = self.get_blob(bucket_name, blob_name)
-        blob.delete()
-        logger.info(f"{blob_name} blob in {bucket_name} bucket deleted.")
+        if blob:
+            blob.delete()
+            logger.info(f"{blob_name} blob in {bucket_name} bucket deleted.")
+        logger.warning(f"Blob: {blob} doesn't exist.")
 
     def upload_table(
         self,
